+++ controlgallery/AtomUIGallery/ShowCases/Other/Splash/ShowCaseControls/GallerySplashWindow.axaml.cs
@@ -1,0 +1,11 @@
+using AtomUI.Desktop.Controls;
+
+namespace AtomUIGallery.ShowCases.Splash;
+
+internal partial class GallerySplashWindow : SplashWindow
+{
+    public GallerySplashWindow()
+    {
+        InitializeComponent();
+    }
+}