+++ controlgallery/AtomUIGallery/ShowCases/Other/Splash/ShowCaseControls/GalleryWindowSplash.axaml.cs
@@ -1,0 +1,13 @@
+using SplashControl = AtomUI.Desktop.Controls.Splash;
+
+namespace AtomUIGallery.ShowCases.Splash;
+
+internal sealed partial class GalleryWindowSplash : SplashControl
+{
+    public GalleryWindowSplash()
+    {
+        InitializeComponent();
+    }
+
+    protected override Type StyleKeyOverride { get; } = typeof(SplashControl);
+}